Added $scope data to the page also added navbar
Added some parts of the event object that is in the $scope of the program. Added some notes to this for future reference.

Also divided things up into bootstrap rows and an image of the angular logo using the ng-src so that we could use the angular `{{}}` notation inside the alt text and image string.