Added filters to allow sorting and a dropdown
Filters were added so that the events can be sorted by name or upvotes and a drop down box has been made to allow users to choose which one it is filtered by.

Some other filters are displayed in this as well but will likely be removed or modified in future commits.